fix:Super Admin validations
